requirements.txt
# 互联网谣言判断系统 - 依赖管理
#
# 版本约束说明：
# - >=X.Y.Z,<X+1.0.0 : 允许补丁和次版本更新，但锁定主版本
# - ==X.Y.Z : 精确版本锁定，用于关键依赖
#
# 依赖版本锁定原因：v0.2.0
# - 问题：原 requirements.txt 未锁定版本，导致部署环境不一致
# - 影响：可能导致行为不一致、部署失败、安全漏洞
# - 解决：使用语义化版本约束，确保稳定性和可维护性

# ========================================
# 核心框架 (LangChain生态)
# ========================================
# LangChain核心 - 允许次版本更新（向后兼容性好）
langchain-core>=0.1.0,<2.0.0
langchain>=0.1.0,<2.0.0

# LangChain集成 - 锁定主版本（API变化较快）
langchain-openai>=0.1.0,<2.0.0
langchain-community>=0.1.0,<1.0.0
langchain-chroma>=0.1.0,<1.0.0
langchain-text-splitters>=0.1.0,<1.0.0
langchain-tavily>=0.1.0,<1.0.0

# ========================================
# Web框架和服务
# ========================================
# FastAPI - 锁定主版本（快速演进中）
fastapi>=0.100.0,<0.200.0
uvicorn>=0.20.0,<1.0.0

# Gradio - 锁定主版本（UI框架变化较快）
gradio>=4.0.0,<7.0.0

# ========================================
# 数据处理和验证
# ========================================
# Pydantic v2 - 锁定主版本（重大变化）
pydantic>=2.0.0,<3.0.0

# Pandas - 锁定主版本
pandas>=2.0.0,<3.0.0

# ========================================
# 存储和缓存
# ========================================
# Diskcache - 锁定主版本
diskcache>=5.0.0,<6.0.0

# ========================================
# HTTP和网络
# ========================================
# Requests - 锁定主版本
requests>=2.31.0,<3.0.0

# DuckDuckGo搜索 - 锁定主版本
ddgs>=9.0.0,<10.0.0

# ========================================
# 可观测性
# ========================================
# 结构化日志
structlog>=24.0.0,<30.0.0

# Prometheus监控
prometheus-client>=0.20.0,<1.0.0

# ========================================
# 工具库
# ========================================
# 环境变量管理
python-dotenv>=1.0.0,<2.0.0

# 重试机制
tenacity>=8.0.0,<10.0.0

# 快速字符串相似度计算 [v0.7.2]
rapidfuzz>=3.0.0,<4.0.0

# 限流（阶段2）
slowapi>=0.1.0,<1.0.0
